Fix: header attribute value in matrixdynamic cells
The header attribute value for a matrixdynamic table cell was not
being populated with a columns title value. This meant that when
a survey is viewed in small screen there is no label for an input.

This fix locates the locTitle.renderedHtml on the cells column within
the getHeaderText method.