Change "Previous" translation for Japanese
We have deployed a survey in Japanese, and received the feedback that the current Japanese default translation of the "Previous" button is incorrect or at least unusual. This PR fixes this and uses the correct term (which was confirmed to us as being correct through several people whose first language is Japanese).